destroyコントローラーにおける条件分岐の追加

+++ app/controllers/items_controller.rb
@@ -37,6 +37,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    unless current_user.id == @item.user_id
+      redirect_to root_path
+    end
+
     if @item.destroy
       redirect_to root_path
     end
